+++ sympy_extras/concrete/__init__.py
@@ -9,9 +9,11 @@ from .qhyper import (QPochhammer, qpochhammer, qbinomial, q_ratio, normal_in, qg
 from .rational import abramov_decomposition, rational_sum, RationalDecomposition
 from .convergence import sum_convergence, product_convergence, is_convergent
 from .dirichlet import dirichlet_series
+from .eulersums import euler_sum, polygamma_series, polygamma_integral_representation
 
 __all__ = ['PiSigmaField', 'Extension', 'karr_term', 'karr_sum', 'summation',
     'build_pisigma_field', 'zeilberger', 'zeilberger_sum', 'wz_certificate', 'wz_prove',
     'Telescoper', 'QPochhammer', 'qpochhammer', 'qbinomial', 'q_ratio', 'normal_in', 'qgosper_term',
     'qgosper_sum', 'qzeilberger', 'QTelescoper', 'abramov_decomposition', 'rational_sum',
-    'RationalDecomposition', 'sum_convergence', 'product_convergence', 'is_convergent', 'dirichlet_series']
+    'RationalDecomposition', 'sum_convergence', 'product_convergence', 'is_convergent', 'dirichlet_series',
+    'euler_sum', 'polygamma_series', 'polygamma_integral_representation']